:lipstick: Move votes download btn inside MesVotes page, improve MesVotes & Resultats header appearance
