+++ frontend/src/assets/styles/pages/_MesVotes.scss
@@ -1,7 +1,15 @@
 .MesVotes {
+  &__header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 1rem;
+  }
+
   &__title {
-    margin-top: 1rem;
-    text-align: center;
+    display: block;
+    padding-top: 0.25rem;
+    padding-bottom: 0.25rem;
   }
 
   &__results {
