+++ frontend/src/pages/About.jsx
@@ -4,7 +4,6 @@ import { Button } from "@mui/material";
 import {
   Delete,
   Email,
-  FileDownload,
   GitHub,
   Instagram,
   Description,
@@ -136,29 +135,6 @@ export default function About() {
             disableElevation
           >
             Réinitialiser mes votes
-          </Button>
-
-          <Button
-            className="About__btn Btn Btn--secondary"
-            startIcon={<FileDownload />}
-            size="large"
-            variant="contained"
-            onClick={() => {
-              const data = new Blob([JSON.stringify(context.choices)], {
-                type: "application/json",
-              });
-              const url = window.URL.createObjectURL(data);
-              const a = document.createElement("a");
-              a.style.display = "none";
-              a.href = url;
-              a.download = "votes_legislatives.txt";
-              document.body.appendChild(a);
-              a.click();
-              window.URL.revokeObjectURL(url);
-            }}
-            disableElevation
-          >
-            Télécharger mes votes
           </Button>
         </div>
